requirements.txt
jupyter
ipykernel
numpy
rdkit-pypi
matplotlib
tqdm
wandb
imageio
networkx
torch-geometric
pytorch_lightning
hydra-core
accelerate
transformers
pandas
matplotlib